Cloud_Eng_Project-main/requirements.txt
boto3==1.34.101
botocore==1.34.101
joblib==1.2.0
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
pytest==7.4.0
PyYAML==6.0.1
scikit_learn==1.2.2
